refactor(svelte): split interaction brush-zoom suite by concern
Close #951. Break the 814-line brush + brush-to-zoom monolith into
tests/interaction/brush-zoom/ (precise-bounds, brush-gesture,
brush-to-zoom). Shared drag helper moves to interaction-harness;
assertion bodies and it-titles are preserved.
